requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
boto3==1.28.32
botocore==1.31.32
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
einops==0.6.1
encodec==0.1.1
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.271
langsmith==0.0.26
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
openai==0.27.9
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
packaging==23.1
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pynput==1.7.6
pyobjc-core==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.12.0
s3transfer==0.6.2
safetensors==0.3.2
scipy==1.11.2
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
stack-data==0.6.2
streamlit==1.28.1
suno-bark @ git+https://github.com/suno-ai/bark.git@56b0ba13f7c281cbffa07ea9abf7b30273a60b6a
sympy==1.12
tdqm==0.0.1
tenacity==8.2.3
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@b413e0610b42d4c8d9c7a69c06440ad27c69808b
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
validators==0.22.0
wcwidth==0.2.6
whisper-mic==1.2.1
yarl==1.9.2
zipp==3.16.2
